+++ backend/src/main/java/kernel360/techpick/feature/link/controller/LinkController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import kernel360.techpick.core.exception.feature.link.ApiLinkException;
+import kernel360.techpick.feature.link.exception.ApiLinkException;
 import kernel360.techpick.feature.link.model.dto.LinkResponse;
 import kernel360.techpick.feature.link.service.LinkService;
 import lombok.RequiredArgsConstructor;
