+++ backend/src/main/java/kernel360/techpick/feature/link/model/provider/LinkProvider.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.springframework.stereotype.Component;
 
-import kernel360.techpick.core.exception.feature.link.ApiLinkException;
+import kernel360.techpick.feature.link.exception.ApiLinkException;
 import kernel360.techpick.core.model.link.Link;
 import kernel360.techpick.feature.link.repository.LinkRepository;
 import lombok.RequiredArgsConstructor;
