+++ backend/src/main/java/kernel360/techpick/feature/link/service/LinkService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import kernel360.techpick.core.exception.feature.link.ApiLinkException;
+import kernel360.techpick.feature.link.exception.ApiLinkException;
 import kernel360.techpick.core.model.link.Link;
 import kernel360.techpick.feature.link.model.dto.LinkRequest;
 import kernel360.techpick.feature.link.model.dto.LinkResponse;
