+++ backend/src/main/java/kernel360/techpick/feature/user/exception/ApiUserErrorCode.java
@@ -1,4 +1,4 @@
-package kernel360.techpick.core.exception.feature.user;
+package kernel360.techpick.feature.user.exception;
 
 import org.springframework.http.HttpStatus;
